Add support for CAN logging (#342)
There are a couple of UI changes that I think most people will appreciate. First, the "start/stop recording" button is much bigger, so it's easier to press it with your thumb while you're driving. :) Second, all of the parameters are now always displayed on the left side of the window.

And, since I have a couple of CAN devices in my Corvette, so I added support for CAN logging. Currently, only one CAN interface is supported. But, it's cheap:

https://www.seeedstudio.com/USB-CAN-Analyzer-p-2888.html

If you need AFR in your logs, I recommend AEM's X-Series gauges, which support support CAN, so you don't need to wire them into your PCM.

There are also a few new log parameters for the 7603 operating system, but I have to admit that not all of them are reliable.

There are also a couple of tiny improvments to PCM Hammer. The long pause before "Kernel upload 48%" always makes my heart skip a beat, so now it prints "Permission to upload granted" before it start the upload. Also, it logs the VPW messages for OBDX devices.